==> 1.scrap.ipynb <==
analysis link:
- https://medium.com/@naviubhi29/data-analysis-and-visualization-on-myanimelist-data-71129f499d7a
- https://id.quora.com/Mengapa-genre-Slice-of-Life-banyak-yg-bilang-puncak-dari-wibu-Bukan-kah-itu-artinya-mulai-bosen-nonton-anime

import requests
from bs4 import BeautifulSoup
import pandas as pd
import re
from PIL import Image
import uuid 
import os
import json
import gc
import time
import traceback

from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.by import By

# Scrap Funtion

# def scrapData(start_ ,end_):
#     for i in range(start_, end_, 50):      

#         file_path_csv = f'./dataset/data/data_{i}.csv'
#         if os.path.exists(file_path_csv):
#             continue
        
#         options = webdriver.ChromeOptions()
#         options.add_argument("--headless")  # Run Chrome in headless mode
#         options.add_argument("--disable-blink-features=AutomationControlled")

#         service = Service(ChromeDriverManager().install())
#         driver = webdriver.Chrome(service=service, options=options)
#         driver_detail = webdriver.Chrome(service=service, options=options)
        
#         driver.get(f"https://myanimelist.net/topanime.php?limit={i}")
        
#         # Tunggu hingga elemen ranking-list muncul
#         try:
#             WebDriverWait(driver, 10).until(
#                 EC.presence_of_element_located((By.CLASS_NAME, "ranking-list"))
#             )
#         except:
#             driver.quit()
#             continue
        
#         page = driver.page_source
#         soup = BeautifulSoup(page, 'html.parser')
#         list = soup.find_all('tr', class_='ranking-list')
        
#         if not list:
#             # print(f"No data found for limit={i}. Skipping...")
#             driver.quit()
#             continue
        
#         data = []

#         for idx, j in enumerate(list):
#             uuid_data = str(uuid.uuid4())
#             title = j.find('div', class_='di-ib').text
#             rank = j.find('span', class_='lightLink').text
#             score = j.find('div', class_='js-top-ranking-score-col').text
#             link = j.find('a').get('href')
#             mal_id = link.split('/')[4]
            
#             anime_data = {
#                 'id': uuid_data,
#                 'mal_id': mal_id,
#                 'title': title,
#                 'image_url': '',
#                 'synopsis': '',
#                 'aired': '',
#                 'premiered': '',
#                 'member': '',
#                 'favorite': '',
#                 'source': '',
#                 'rank': rank,
#                 'link': link,
#                 'episode': 0,
#                 'type': '',
#                 'genre': '',
#                 'producer': '',
#                 'studio': '',
#                 'theme': '',
#                 'demographic': '',
#                 'duration': '',
#                 'rating': '',
#                 'score': score,
#                 'user_score': '',
#             }
            
#             # # Detail 
#             driver_detail.get(link)
#             page_detail = driver_detail.page_source
#             soup_detail = BeautifulSoup(page_detail, 'html.parser')

#             # Image
#             image_url_ = soup_detail.find('img', itemprop='image')
#             if image_url_:
#                 anime_data['image_url'] =  image_url_['data-src']
            
            
#             # Genre
#             genre_list = soup_detail.find_all('span', itemprop='genre')
#             if genre_list:
#                 genres = [g.text for g in genre_list]
#                 anime_data['genre'] = ", ".join(genres)
            
#             # Synopsis
#             synopsis_ = soup_detail.find('p', itemprop='description')
#             if synopsis_:
#                 cleaned_synopsis = synopsis_.text.replace('\n', ' ')
#                 anime_data['synopsis'] = cleaned_synopsis
                
#             # Producer
#             producer_ = soup_detail.find('span', class_='dark_text', text='Producers:')
#             if producer_:
#                 producers = [a.text for a in producer_.find_next_siblings('a')]
#                 anime_data['producer'] = ", ".join(producers)
            
#             # Studio
#             studio_ = soup_detail.find('span', class_='dark_text', text='Studios:')
#             if studio_:
#                 studios = [a.text for a in studio_.find_next_siblings('a')]
#                 anime_data['studio'] = ", ".join(studios)

#             # Theme
#             theme_ = soup_detail.find('span', class_='dark_text', text='Theme:')
#             if theme_:
#                 theme_ = [a.text for a in theme_.find_next_siblings('a')]
#                 anime_data['theme'] = ", ".join(theme_)

#             # Demographic
#             demographic_ = soup_detail.find('span', class_='dark_text', text='Demographic:')
#             if demographic_:
#                 demographic_ = [a.text for a in demographic_.find_next_siblings('a')]
#                 anime_data['demographic'] = ", ".join(demographic_)

#             # Episode
#             episode_ = soup_detail.find('span', class_='dark_text', text='Episodes:')
#             if episode_:
#                 anime_data['episode'] = episode_.find_parent().text.strip().split(':', 1)[1].replace("\n", "").replace(" ", "")

#             # Type
#             type_ = soup_detail.find('span', class_='dark_text', text='Type:')
#             if type_:
#                 anime_data['type'] = type_.find_parent().text.strip().split(':', 1)[1].replace("\n", "").replace(" ", "")

#             # Duration
#             duration_ = soup_detail.find('span', class_='dark_text', text='Duration:')
#             if duration_:
#                 anime_data['duration'] = duration_.find_parent().text.strip().split(':', 1)[1].replace("\n", "")

#             # Rating
#             rating_ = soup_detail.find('span', class_='dark_text', text='Rating:')
#             if rating_:
#                 anime_data['rating'] = rating_.find_parent().text.strip().split(':', 1)[1].replace("\n", "")
                
#             # Aired
#             aired_ = soup_detail.find('span', class_='dark_text', text='Aired:')
#             if aired_:
#                 anime_data['aired'] = aired_.find_parent().text.strip().split(':', 1)[1].replace("\n", "")
                
#             # Premiered
#             premiered_ = soup_detail.find('span', class_='dark_text', text='Premiered:')
#             if premiered_:
#                 anime_data['premiered'] = premiered_.find_parent().text.strip().split(':', 1)[1].replace("\n", "")

#             # Source
#             source_ = soup_detail.find('span', class_='dark_text', text='Source:')
#             if source_:
#                 anime_data['source'] = source_.find_parent().text.strip().split(':', 1)[1].replace("\n", "")

#             # Member
#             member_ = soup_detail.find('span', class_='dark_text', text='Members:')
#             if member_:
#                 anime_data['member'] = member_.find_parent().text.strip().split(':', 1)[1].replace("\n", "").replace(",", "")

#             # Favorite
#             favorite_ = soup_detail.find('span', class_='dark_text', text='Favorites:')
#             if favorite_:
#                 anime_data['favorite'] = favorite_.find_parent().text.strip().split(':', 1)[1].replace("\n", "").replace(",", "")
                
#             # User vote the Score
#             user_score_ = soup_detail.find('span', itemprop='ratingCount')
#             if user_score_:
#                 anime_data['user_score'] =  user_score_.text

#             data.append(anime_data) 

#             # Print progress
#             progress = (idx + 1) / len(list) * 100
#             print(f"\rScraping Data ke {i} [{'=' * int(progress // 10)}{' ' * (10 - int(progress // 10))}] {idx + 1}/{len(list)}", end="", flush=True)
            
#         df = pd.DataFrame(data)
#         df.to_csv(file_path_csv, index=False)
        
#         gc.collect()
        
#         print()
#         # driver.quit()
    
#     return print(f"Data Scrap {start_} to {end_} Completed")

def scrapData(start_ ,end_):
    for i in range(start_, end_, 50):      

        file_path_csv = f'./dataset/data/data_{i}.csv'
        if os.path.exists(file_path_csv):
            continue
        
        attempt = 0
        success = False
        error_tolerance = 3
        
        while attempt < error_tolerance and not success:
            try:
                options = webdriver.ChromeOptions()
                options.add_argument("--headless")  # Run Chrome in headless mode
                options.add_argument("--disable-blink-features=AutomationControlled")

                service = Service(ChromeDriverManager().install())
                driver = webdriver.Chrome(service=service, options=options)
                driver_detail = webdriver.Chrome(service=service, options=options)
                
                driver.get(f"https://myanimelist.net/topanime.php?limit={i}")
                
                # Tunggu hingga elemen ranking-list muncul
                try:
                    WebDriverWait(driver, 10).until(
                        EC.presence_of_element_located((By.CLASS_NAME, "ranking-list"))
                    )
                except:
                    driver.quit()
                    continue

                
                page = driver.page_source
                soup = BeautifulSoup(page, 'html.parser')
                list = soup.find_all('tr', class_='ranking-list')
                
                if not list:
                    driver.quit()
                    continue
                
                data = []

                for idx, j in enumerate(list):
                    uuid_data = str(uuid.uuid4())
                    title = j.find('div', class_='di-ib').text
                    rank = j.find('span', class_='lightLink').text
                    score = j.find('div', class_='js-top-ranking-score-col').text
                    link = j.find('a').get('href')
                    mal_id = link.split('/')[4]
                    
                    anime_data = {
                        'id': uuid_data,
                        'mal_id': mal_id,
                        'title': title,
                        'image_url': '',
                        'synopsis': '',
                        'aired': '',
                        'premiered': '',
                        'member': '',
                        'favorite': '',
                        'source': '',
                        'rank': rank,
                        'link': link,
                        'episode': 0,
                        'type': '',
                        'genre': '',
                        'producer': '',
                        'studio': '',
                        'theme': '',
                        'demographic': '',
                        'duration': '',
                        'rating': '',
                        'score': score,
                        'user_score': '',
                    }
                    
                    # # Detail 
                    driver_detail.get(link)
                    page_detail = driver_detail.page_source
                    soup_detail = BeautifulSoup(page_detail, 'html.parser')

                    # Image
                    image_url_ = soup_detail.find('img', itemprop='image')
                    if image_url_:
                        anime_data['image_url'] =  image_url_['data-src']
                    
                    
                    # Genre
                    genre_list = soup_detail.find_all('span', itemprop='genre')
                    if genre_list:
                        genres = [g.text for g in genre_list]
                        anime_data['genre'] = ", ".join(genres)
                    
                    # Synopsis
                    synopsis_ = soup_detail.find('p', itemprop='description')
                    if synopsis_:
                        cleaned_synopsis = synopsis_.text.replace('\n', ' ')
                        anime_data['synopsis'] = cleaned_synopsis
                        
                    # Producer
                    producer_ = soup_detail.find('span', class_='dark_text', text='Producers:')
                    if producer_:
                        producers = [a.text for a in producer_.find_next_siblings('a')]
                        anime_data['producer'] = ", ".join(producers)
                    
                    # Studio
                    studio_ = soup_detail.find('span', class_='dark_text', text='Studios:')
                    if studio_:
                        studios = [a.text for a in studio_.find_next_siblings('a')]
                        anime_data['studio'] = ", ".join(studios)

                    # Theme
                    theme_ = soup_detail.find('span', class_='dark_text', text='Theme:')
                    if theme_:
                        theme_ = [a.text for a in theme_.find_next_siblings('a')]
                        anime_data['theme'] = ", ".join(theme_)

                    # Demographic
                    demographic_ = soup_detail.find('span', class_='dark_text', text='Demographic:')
                    if demographic_:
                        demographic_ = [a.text for a in demographic_.find_next_siblings('a')]
                        anime_data['demographic'] = ", ".join(demographic_)

                    # Episode
                    episode_ = soup_detail.find('span', class_='dark_text', text='Episodes:')
                    if episode_:
                        anime_data['episode'] = episode_.find_parent().text.strip().split(':', 1)[1].replace("\n", "").replace(" ", "")

                    # Type
                    type_ = soup_detail.find('span', class_='dark_text', text='Type:')
                    if type_:
                        anime_data['type'] = type_.find_parent().text.strip().split(':', 1)[1].replace("\n", "").replace(" ", "")

                    # Duration
                    duration_ = soup_detail.find('span', class_='dark_text', text='Duration:')
                    if duration_:
                        anime_data['duration'] = duration_.find_parent().text.strip().split(':', 1)[1].replace("\n", "")

                    # Rating
                    rating_ = soup_detail.find('span', class_='dark_text', text='Rating:')
                    if rating_:
                        anime_data['rating'] = rating_.find_parent().text.strip().split(':', 1)[1].replace("\n", "")
                        
                    # Aired
                    aired_ = soup_detail.find('span', class_='dark_text', text='Aired:')
                    if aired_:
                        anime_data['aired'] = aired_.find_parent().text.strip().split(':', 1)[1].replace("\n", "")
                        
                    # Premiered
                    premiered_ = soup_detail.find('span', class_='dark_text', text='Premiered:')
                    if premiered_:
                        anime_data['premiered'] = premiered_.find_parent().text.strip().split(':', 1)[1].replace("\n", "")

                    # Source
                    source_ = soup_detail.find('span', class_='dark_text', text='Source:')
                    if source_:
                        anime_data['source'] = source_.find_parent().text.strip().split(':', 1)[1].replace("\n", "")

                    # Member
                    member_ = soup_detail.find('span', class_='dark_text', text='Members:')
                    if member_:
                        anime_data['member'] = member_.find_parent().text.strip().split(':', 1)[1].replace("\n", "").replace(",", "")

                    # Favorite
                    favorite_ = soup_detail.find('span', class_='dark_text', text='Favorites:')
                    if favorite_:
                        anime_data['favorite'] = favorite_.find_parent().text.strip().split(':', 1)[1].replace("\n", "").replace(",", "")
                        
                    # User vote the Score
                    user_score_ = soup_detail.find('span', itemprop='ratingCount')
                    if user_score_:
                        anime_data['user_score'] =  user_score_.text

                    data.append(anime_data) 

                    # Print progress
                    progress = (idx + 1) / len(list) * 100
                    print(f"\rScraping Data ke {i} [{'=' * int(progress // 10)}{' ' * (10 - int(progress // 10))}] {idx + 1}/{len(list)}", end="", flush=True)
                    
                df = pd.DataFrame(data)
                df.to_csv(file_path_csv, index=False)
                
                gc.collect()
                print()
                success = True
                
            except Exception as e:
                attempt += 1
                time.sleep(5)  # jeda sebelum retry
            finally:
                try:
                    driver.quit()
                    driver_detail.quit()
                except:
                    pass
    
    return print(f"Data Scrap {start_} to {end_} Completed")

## Start Scrap Round 2
Start at _**12 August 2025**_ 

Finish at _**15 August 2025**_

scrapData(0, 5050)

scrapData(5050, 10050)

scrapData(10050, 15050)

scrapData(15050, 20050)

scrapData(20050, 25050)

scrapData(25050, 28750)
