==> 5.scrap-comment.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import re
from PIL import Image
import uuid  # Import UUID module
import os
import json

from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.by import By

import pandas as pd

# Nisekoi

options = webdriver.ChromeOptions()
options.add_argument("--headless")
options.add_argument("--disable-blink-features=AutomationControlled")

service = Service(ChromeDriverManager().install())
driver = webdriver.Chrome(service=service, options=options)
driver_detail = webdriver.Chrome(service=service, options=options)

animeId = 18897
anime_name = "Nisekoi"

pages = 1
to_pages = 100

data = []
seen_reviews = set()

for i in range(pages, to_pages + 1):
    link = f"https://myanimelist.net/anime/{animeId}/{anime_name}/reviews?sort=suggested&filter_check=&filter_hide=&preliminary=on&spoiler=off&p={i}"
    driver.get(link)

    try:
        WebDriverWait(driver, 10).until(
            EC.presence_of_element_located((By.CLASS_NAME, "review-element"))
        )
    except:
        break


    page = driver.page_source
    soup = BeautifulSoup(page, 'html.parser')
    list = soup.find_all('div', class_='review-element')

    for idx, j in enumerate(list):
        
        username = [div.text for div in j.find_all('div', class_='username')][0].strip().replace('★', '')
        update_at = [div.text for div in j.find_all('div', class_='update_at')][0].strip().replace('★', '')
        
        duplicate_check = f"{username} - {update_at}"
        
        if duplicate_check in seen_reviews:
            print(f"Data duplikat ditemukan di halaman {i}. Menghentikan scraping.")
            break  # Keluar dari loop review
        
        seen_reviews.add(duplicate_check)  # Tambahkan ke set data unik
        
        comment = [div.text for div in j.find_all('div', class_='text')][0].strip().replace('★', '')
        tags = j.find('div', class_='tags').find('div', class_='tag').text
        rating = [div.text for div in j.find_all('div', class_='rating')][0].split(':')[1].strip().replace('★', '')
        icon_reaction = [div.text for div in j.find_all('div', class_='icon-reaction')][0].strip().replace('★', '')
        
        review_data = {
            'username': username,
            'update_at': update_at,
            'tags': tags,
            'comment': comment,
            'reaction': icon_reaction,
            'rating': rating
        }
        
        data.append(review_data)
    else: 
        continue
    
    break

df = pd.DataFrame(data)
df

df.to_csv('nisekoi/Nisekoi.csv', index=False)

# Nisekoi_

options = webdriver.ChromeOptions()
options.add_argument("--headless")
options.add_argument("--disable-blink-features=AutomationControlled")

service = Service(ChromeDriverManager().install())
driver = webdriver.Chrome(service=service, options=options)
driver_detail = webdriver.Chrome(service=service, options=options)


animeId = 27787
anime_name = "Nisekoi_"

pages = 1
to_pages = 100

data = []
seen_reviews = set()

for i in range(pages, to_pages + 1):
    link = f"https://myanimelist.net/anime/{animeId}/{anime_name}/reviews?sort=suggested&filter_check=&filter_hide=&preliminary=on&spoiler=off&p={i}"
    driver.get(link)

    try:
        WebDriverWait(driver, 10).until(
            EC.presence_of_element_located((By.CLASS_NAME, "review-element"))
        )
    except:
        break


    page = driver.page_source
    soup = BeautifulSoup(page, 'html.parser')
    list = soup.find_all('div', class_='review-element')

    for idx, j in enumerate(list):
        
        username = [div.text for div in j.find_all('div', class_='username')][0].strip().replace('★', '')
        update_at = [div.text for div in j.find_all('div', class_='update_at')][0].strip().replace('★', '')
        
        duplicate_check = f"{username} - {update_at}"
        
        if duplicate_check in seen_reviews:
            print(f"Data duplikat ditemukan di halaman {i}. Menghentikan scraping.")
            break  # Keluar dari loop review
        
        seen_reviews.add(duplicate_check)  # Tambahkan ke set data unik
        
        comment = [div.text for div in j.find_all('div', class_='text')][0].strip().replace('★', '')
        tags = j.find('div', class_='tags').find('div', class_='tag').text
        rating = [div.text for div in j.find_all('div', class_='rating')][0].split(':')[1].strip().replace('★', '')
        icon_reaction = [div.text for div in j.find_all('div', class_='icon-reaction')][0].strip().replace('★', '')
        
        review_data = {
            'username': username,
            'update_at': update_at,
            'tags': tags,
            'comment': comment,
            'reaction': icon_reaction,
            'rating': rating
        }
        
        data.append(review_data)
    else: 
        continue
    
    break

df = pd.DataFrame(data)
df

df.to_csv('nisekoi/Nisekoi_.csv', index=False)

# Nisekoi_OVA

options = webdriver.ChromeOptions()
options.add_argument("--headless")
options.add_argument("--disable-blink-features=AutomationControlled")

service = Service(ChromeDriverManager().install())
driver = webdriver.Chrome(service=service, options=options)
driver_detail = webdriver.Chrome(service=service, options=options)

animeId = 24227
anime_name = "Nisekoi_OVA"

pages = 1
to_pages = 100

data = []
seen_reviews = set()

for i in range(pages, to_pages + 1):
    link = f"https://myanimelist.net/anime/{animeId}/{anime_name}/reviews?sort=suggested&filter_check=&filter_hide=&preliminary=on&spoiler=off&p={i}"
    driver.get(link)

    try:
        WebDriverWait(driver, 10).until(
            EC.presence_of_element_located((By.CLASS_NAME, "review-element"))
        )
    except:
        break


    page = driver.page_source
    soup = BeautifulSoup(page, 'html.parser')
    list = soup.find_all('div', class_='review-element')

    for idx, j in enumerate(list):
        
        username = [div.text for div in j.find_all('div', class_='username')][0].strip().replace('★', '')
        update_at = [div.text for div in j.find_all('div', class_='update_at')][0].strip().replace('★', '')
        
        duplicate_check = f"{username} - {update_at}"
        
        if duplicate_check in seen_reviews:
            print(f"Data duplikat ditemukan di halaman {i}. Menghentikan scraping.")
            break  # Keluar dari loop review
        
        seen_reviews.add(duplicate_check)  # Tambahkan ke set data unik
        
        comment = [div.text for div in j.find_all('div', class_='text')][0].strip().replace('★', '')
        tags = j.find('div', class_='tags').find('div', class_='tag').text
        rating = [div.text for div in j.find_all('div', class_='rating')][0].split(':')[1].strip().replace('★', '')
        icon_reaction = [div.text for div in j.find_all('div', class_='icon-reaction')][0].strip().replace('★', '')
        
        review_data = {
            'username': username,
            'update_at': update_at,
            'tags': tags,
            'comment': comment,
            'reaction': icon_reaction,
            'rating': rating
        }
        
        data.append(review_data)
    else: 
        continue
    
    break

df = pd.DataFrame(data)
df

df.to_csv('nisekoi/Nisekoi_OVA.csv', index=False)

# Nisekoi__OVA

options = webdriver.ChromeOptions()
options.add_argument("--headless")
options.add_argument("--disable-blink-features=AutomationControlled")

service = Service(ChromeDriverManager().install())
driver = webdriver.Chrome(service=service, options=options)
driver_detail = webdriver.Chrome(service=service, options=options)

animeId = 32553
anime_name = "Nisekoi__OVA"

pages = 1
to_pages = 100

data = []
seen_reviews = set()

for i in range(pages, to_pages + 1):
    link = f"https://myanimelist.net/anime/{animeId}/{anime_name}/reviews?sort=suggested&filter_check=&filter_hide=&preliminary=on&spoiler=off&p={i}"
    driver.get(link)

    try:
        WebDriverWait(driver, 10).until(
            EC.presence_of_element_located((By.CLASS_NAME, "review-element"))
        )
    except:
        break


    page = driver.page_source
    soup = BeautifulSoup(page, 'html.parser')
    list = soup.find_all('div', class_='review-element')

    for idx, j in enumerate(list):
        
        username = [div.text for div in j.find_all('div', class_='username')][0].strip().replace('★', '')
        update_at = [div.text for div in j.find_all('div', class_='update_at')][0].strip().replace('★', '')
        
        duplicate_check = f"{username} - {update_at}"
        
        if duplicate_check in seen_reviews:
            print(f"Data duplikat ditemukan di halaman {i}. Menghentikan scraping.")
            break  # Keluar dari loop review
        
        seen_reviews.add(duplicate_check)  # Tambahkan ke set data unik
        
        comment = [div.text for div in j.find_all('div', class_='text')][0].strip().replace('★', '')
        tags = j.find('div', class_='tags').find('div', class_='tag').text
        rating = [div.text for div in j.find_all('div', class_='rating')][0].split(':')[1].strip().replace('★', '')
        icon_reaction = [div.text for div in j.find_all('div', class_='icon-reaction')][0].strip().replace('★', '')
        
        review_data = {
            'username': username,
            'update_at': update_at,
            'tags': tags,
            'comment': comment,
            'reaction': icon_reaction,
            'rating': rating
        }
        
        data.append(review_data)
    else: 
        continue
    
    break

df = pd.DataFrame(data)
df

df.to_csv('nisekoi/Nisekoi__OVA.csv', index=False)

# Nisekoimonogatari

options = webdriver.ChromeOptions()
options.add_argument("--headless")
options.add_argument("--disable-blink-features=AutomationControlled")

service = Service(ChromeDriverManager().install())
driver = webdriver.Chrome(service=service, options=options)
driver_detail = webdriver.Chrome(service=service, options=options)

animeId = 30514
anime_name = "Nisekoimonogatari"

pages = 1
to_pages = 100

data = []
seen_reviews = set()

for i in range(pages, to_pages + 1):
    link = f"https://myanimelist.net/anime/{animeId}/{anime_name}/reviews?sort=suggested&filter_check=&filter_hide=&preliminary=on&spoiler=off&p={i}"
    driver.get(link)

    try:
        WebDriverWait(driver, 10).until(
            EC.presence_of_element_located((By.CLASS_NAME, "review-element"))
        )
    except:
        break


    page = driver.page_source
    soup = BeautifulSoup(page, 'html.parser')
    list = soup.find_all('div', class_='review-element')

    for idx, j in enumerate(list):
        
        username = [div.text for div in j.find_all('div', class_='username')][0].strip().replace('★', '')
        update_at = [div.text for div in j.find_all('div', class_='update_at')][0].strip().replace('★', '')
        
        duplicate_check = f"{username} - {update_at}"
        
        if duplicate_check in seen_reviews:
            print(f"Data duplikat ditemukan di halaman {i}. Menghentikan scraping.")
            break  # Keluar dari loop review
        
        seen_reviews.add(duplicate_check)  # Tambahkan ke set data unik
        
        comment = [div.text for div in j.find_all('div', class_='text')][0].strip().replace('★', '')
        tags = j.find('div', class_='tags').find('div', class_='tag').text
        rating = [div.text for div in j.find_all('div', class_='rating')][0].split(':')[1].strip().replace('★', '')
        icon_reaction = [div.text for div in j.find_all('div', class_='icon-reaction')][0].strip().replace('★', '')
        
        review_data = {
            'username': username,
            'update_at': update_at,
            'tags': tags,
            'comment': comment,
            'reaction': icon_reaction,
            'rating': rating
        }
        
        data.append(review_data)
    else: 
        continue
    
    break

df = pd.DataFrame(data)
df

df.to_csv('nisekoi/Nisekoi_monogatari.csv', index=False)
